#50 Rename package (astro_ph → arxiv_post)
