+++ arxiv_post/detex.py
