+++ docs/build
@@ -6,7 +6,7 @@ pandoc --column 500 -f markdown -t rst -o docs/readme.rst README.md
 
 
 # auto-generate package guide
-sphinx-apidoc -efT -o docs/_apidoc astro_ph
+sphinx-apidoc -efT -o docs/_apidoc arxiv_post
 
 
 # build website
