+++ pyproject.toml
@@ -28,7 +28,7 @@ pytest = "^6.2"
 sphinx = "^4.2"
 
 [tool.poetry.scripts]
-astro-ph = 'astro_ph.cli:cli'
+astro-ph = "arxiv_post.cli:cli"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
